Refactor value package files
